Version 3.0.6
The rounding of rain rate when doing the conversion to in/hr was
rounding too soon and as a result, it would take a very heavy rate
to register at all.  Changed the calculation around so that the
rounding happens at the very end.  Also make sure to set the correct
uom for the rain rate.