Separate out rain query from station uom query
While the query we use to get the station uom does included the
federated observations that include today's and yesterday's
preciptiation.  Those values are only for the the device that is
designaged primary.

If the station has more than one device capabile of collecting rain
data, we need to process each one of those devices separately.